was able to split up and put the server and store pages into their own folders
